fix: fixing the organization of the blog context and deleting the separated crud archive
